Refactor ConfigurationManagerProvider and FileProvider as plugins.
