+++ Reflection.sln
@@ -3,7 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33516.290
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp", "ConsoleApp\ConsoleApp.csproj", "{1CCABB44-FE0D-4F43-925F-6052D1256025}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp", "ConsoleApp\ConsoleApp.csproj", "{1CCABB44-FE0D-4F43-925F-6052D1256025}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileConfigurationProviderPlugin", "FileConfigurationProviderPlugin\FileConfigurationProviderPlugin.csproj", "{CA3FDB23-58CA-4BF4-BDAA-DDAB890E71D2}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConfigurationManagerConfigurationProviderPlugin", "ConfigurationManagerConfigurationProviderPlugin\ConfigurationManagerConfigurationProviderPlugin.csproj", "{DE30EE6D-2153-4715-9D64-3C0B5FAE0CFE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "Common\Common.csproj", "{A5929E1C-650D-4854-847E-13147CFEB418}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +21,18 @@ Global
 		{1CCABB44-FE0D-4F43-925F-6052D1256025}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1CCABB44-FE0D-4F43-925F-6052D1256025}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1CCABB44-FE0D-4F43-925F-6052D1256025}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CA3FDB23-58CA-4BF4-BDAA-DDAB890E71D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA3FDB23-58CA-4BF4-BDAA-DDAB890E71D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA3FDB23-58CA-4BF4-BDAA-DDAB890E71D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA3FDB23-58CA-4BF4-BDAA-DDAB890E71D2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE30EE6D-2153-4715-9D64-3C0B5FAE0CFE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE30EE6D-2153-4715-9D64-3C0B5FAE0CFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE30EE6D-2153-4715-9D64-3C0B5FAE0CFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE30EE6D-2153-4715-9D64-3C0B5FAE0CFE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A5929E1C-650D-4854-847E-13147CFEB418}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A5929E1C-650D-4854-847E-13147CFEB418}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A5929E1C-650D-4854-847E-13147CFEB418}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A5929E1C-650D-4854-847E-13147CFEB418}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
